clang: added .clang-format (IEM plugin style), fixed clang-tidy warnings, updated formatting

+++ src/util.h
@@ -4,14 +4,12 @@
 #include <imgui.h>
 
 #include <array>
-#include <cmath>
-#include <numbers>
 #include <string>
 
 #define DRAW_DEBUG false
 #define FONT "Sono-SemiBold.ttf"
 
-constexpr const char *TITLE = "gemuPG";
+constexpr const char* TITLE = "gemuPG";
 constexpr bool ADJUST_AMP_BY_CREST = true;
 constexpr float ONE_DIV_SQRT_THREE = 0.5773502691896258;
 
@@ -22,7 +20,7 @@ constexpr unsigned int MIN_FPS = 30;
 constexpr double MAX_BPM_DEVIATION = 0.1;
 
 constexpr unsigned int ICON_SIZE = 64;
-constexpr ImVec2 IMGUI_WIN_SIZE = {ICON_SIZE * 8, ICON_SIZE * 5};
+constexpr ImVec2 IMGUI_WIN_SIZE = { ICON_SIZE * 8, ICON_SIZE * 5 };
 constexpr float TEXT_TYPE_SIZE_DIV = 4;
 constexpr float TEXT_VALUE_SIZE_DIV = 2;
 
@@ -32,13 +30,13 @@ constexpr unsigned int TICKS_PER_BAR =
 
 constexpr bool SEQUENCER_RANDOMIZED = true;
 
-constexpr SDL_Color GENERATOR_COLOR = {0, 0, 255, SDL_ALPHA_OPAQUE};
-constexpr SDL_Color AREA_COLOR = {128, 128, 128, SDL_ALPHA_OPAQUE};
-constexpr SDL_Color SEQUENCER_COLOR = {255, 255, 0, SDL_ALPHA_OPAQUE};
+constexpr SDL_Color GENERATOR_COLOR = { 0, 0, 255, SDL_ALPHA_OPAQUE };
+constexpr SDL_Color AREA_COLOR = { 128, 128, 128, SDL_ALPHA_OPAQUE };
+constexpr SDL_Color SEQUENCER_COLOR = { 255, 255, 0, SDL_ALPHA_OPAQUE };
 
 constexpr float BLOCK_SIZE_FACTOR = 0.8f;
 constexpr int GRID_SIZE = 512;
-constexpr SDL_Color GRID_COLOR = {128, 128, 128, SDL_ALPHA_OPAQUE};
+constexpr SDL_Color GRID_COLOR = { 128, 128, 128, SDL_ALPHA_OPAQUE };
 
 constexpr float MIN_CAMERA_ZOOM = 16.0f;
 constexpr float MAX_CAMERA_ZOOM = 512.0f;
@@ -53,78 +51,75 @@ constexpr double ENV_TIME_CONST = 3;
 constexpr double MIN_ENV_TIME_S = 5e-3; // 5ms
 constexpr Uint64 MIN_ENV_SAMPLES = 10;
 
-constexpr double AUDIO_S16_PEAK = 65536 / 2;
-constexpr double AUDIO_S32_PEAK = 4294967296 / 2;
+constexpr double AUDIO_S16_PEAK = static_cast<double> (65536) / 2;
+constexpr double AUDIO_S32_PEAK = static_cast<double> (4294967296) / 2;
 constexpr double AUDIO_F32_PEAK = 1.0f;
 
 using pitch_t = float;
 
-typedef enum {
-  WAVE_SAMPLE,
-  WAVE_SAW,
-  WAVE_SINE,
-  WAVE_SQUARE,
-  WAVE_TRIANGLE
+typedef enum
+{
+    WAVE_SAMPLE,
+    WAVE_SAW,
+    WAVE_SINE,
+    WAVE_SQUARE,
+    WAVE_TRIANGLE
 } WAVE_FORMS;
 
-typedef enum {
-  PITCH_REL_FREQUENCY,
-  PITCH_ABS_FREQUENCY,
-  PITCH_INTERVAL,
-  PITCH_NOTE
+typedef enum
+{
+    PITCH_REL_FREQUENCY,
+    PITCH_ABS_FREQUENCY,
+    PITCH_INTERVAL,
+    PITCH_NOTE
 } pitch_type_t;
 
-typedef enum {
-  AREA,
-  BLOCK_GENERATOR,
-  BLOCK_SEQUENCER,
-  // BLOCK_SETTINGS,
-  // BLOCK_EFFECT,
-  // BLOCK_MODULATOR
+typedef enum
+{
+    AREA,
+    BLOCK_GENERATOR,
+    BLOCK_SEQUENCER,
+    // BLOCK_SETTINGS,
+    // BLOCK_EFFECT,
+    // BLOCK_MODULATOR
 } block_type_t;
 
-template <typename T> struct Vector2 {
-  T x{};
-  T y{};
+template <typename T>
+struct Vector2
+{
+    T x {};
+    T y {};
 
-  constexpr Vector2() = default;
+    constexpr Vector2() = default;
 
-  constexpr Vector2(T x, T y) : x(x), y(y) {}
+    constexpr Vector2 (T x, T y) : x (x), y (y) {}
 
-  Vector2 operator/(T scalar) const { return Vector2(x / scalar, y / scalar); }
-  Vector2 operator+(const Vector2 &other) const {
-    return Vector2(x + other.x, y + other.y);
-  }
-  Vector2 operator-(const Vector2 &other) const {
-    return Vector2(x - other.x, y - other.y);
-  }
-  Vector2 operator*(T scalar) const { return Vector2(x * scalar, y * scalar); }
-  bool operator==(const Vector2 &other) const {
-    return x == other.x && y == other.y;
-  }
-  Vector2 operator+=(const Vector2 &other) const {
-    return {x + other.x, y + other.y};
-  }
+    Vector2 operator/ (T scalar) const { return Vector2 (x / scalar, y / scalar); }
+    Vector2 operator+ (const Vector2& other) const { return Vector2 (x + other.x, y + other.y); }
+    Vector2 operator- (const Vector2& other) const { return Vector2 (x - other.x, y - other.y); }
+    Vector2 operator* (T scalar) const { return Vector2 (x * scalar, y * scalar); }
+    bool operator== (const Vector2& other) const { return x == other.x && y == other.y; }
+    Vector2 operator+= (const Vector2& other) const { return { x + other.x, y + other.y }; }
 };
 
 using Vector2f = Vector2<float>;
 using Vector2i = Vector2<int>;
 
-Vector2f floorVec(Vector2f vec);
+Vector2f floorVec (Vector2f vec);
 
-float calcCrest(std::array<float, WAVE_SIZE> wave);
+float calcCrest (std::array<float, WAVE_SIZE> wave);
 
-ImVec4 toImVec4(SDL_Color color);
+ImVec4 toImVec4 (SDL_Color color);
 
-std::array<Vector2f, 4> getAdjacentPositions(Vector2f pos);
+std::array<Vector2f, 4> getAdjacentPositions (Vector2f pos);
 
-std::pair<std::string, int> freqToNote(pitch_t);
-pitch_t noteToFreq(std::string, int);
-pitch_t intervalToRatio(float interval, float octave_subdivision);
+std::pair<std::string, int> freqToNote (pitch_t);
+pitch_t noteToFreq (std::string, int);
+pitch_t intervalToRatio (float interval, float octave_subdivision);
 
-SDL_Color invertColor(SDL_Color);
+SDL_Color invertColor (SDL_Color);
 
-double interpTable(float *array, size_t size, double idx);
+double interpTable (float* array, size_t size, double idx);
 
 extern float RENDER_SCALE;
 extern Uint32 PERFORMANCE_FREQUENCY;
